requirements.txt
pandas
numpy
tensorflow
scikit-learn
supabase
schedule
python-dotenv